#10: Remove items from Checkout
Added the ability to remove items from the checkout, and changed the way that the total due was being calculated. Instead of managing the addition/subtraction manually through a stored value in Firebase, the total is now a component level computed property that is a product of the Vuex `checkout.items` state.